why would pos ever be negative?

+++ llarp/dnsc.cpp
@@ -299,7 +299,7 @@ generic_handle_dnsc_recvfrom(dnsc_answer_request *request,
       llarp::LogDebug("Read an authority for ",
                      request->question.name, " at ", std::to_string(pos));
     // castBuf += answer->name.length() + 4 + 4 + 4 + answer->rdLen;
-      if(pos > sz)
+      if((ssize_t)pos > sz)
       {
           llarp::LogWarn("Would read past end of dns packet. for ",
                          request->question.name);
